Treat CR + LF equally
Because PSX on Windows seems to emit CR+LF pairs rather than LF alone, this was causing trouble with string interpretation in cdu.

To work around this, we'll treat CR precisely the same as a LF, and won't bother trying to interpret empty lines such as the one that'll be generated inbetween a CR+LF.  Unless PSX starts emiting CRs in its message, with CR+LF as the strict statement terminator, this shouldn't need to change either.